Ejercicios 07 y 08 corregidos
